requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==41.0.3
cssbeautifier==1.14.9
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.55.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djlint==1.32.1
EditorConfig==0.12.3
factory-boy==3.3.0
Faker==19.6.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.9
json5==0.9.14
oauthlib==3.2.2
pathspec==0.11.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
